experiments/requirements_clf.txt
torch==2.0.0
torchvision==0.15.1

timm==1.0.15

numpy==1.24.4
Pillow==11.2.1
opencv-python==4.11.0.86

pandas==2.2.3
tqdm==4.67.1

scikit-learn==1.3.0
faiss-cpu==1.7.3